[refactor] command_buffer directory 생성
