+++ src/command_buffer/command_buffer_optimizer.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod, ABC
 
-from src.command_buffer_data import ERASE_VALUE, ERASE, WRITE, EMPTY, CommandBufferData, MAX_SIZE_OF_COMMAND_BUFFERS, \
+from src.command_buffer.command_buffer_data import ERASE, WRITE, EMPTY, CommandBufferData, MAX_SIZE_OF_COMMAND_BUFFERS, \
     ERASE_CHUNK_SIZE
